API_Detection/requirements.txt
pip
pillow 
numpy
fastapi
uvicorn 
python-multipart